Store symbol and subsection ids as u32
SymbolId and InputSectionId were usize. Every vector of them - each
object's nlist-to-symbol map (4.6M entries on a debug clang link), the
output sections' member lists and each object's subsection list (4.4M
each), the stub/GOT/thread-pointer slot lists, the Option<SymbolId> in
every output symbol table entry - carried 8 bytes per id. mold-rust's
ObjId, SymbolId and InputSectionId are u32 newtypes; make ours u32 too.

To keep the change to the storage width rather than a rewrite of every
arena access, ctx.isecs becomes a thin InputSections wrapper indexable by u32
or usize (through Deref to the Vec), SymbolTable indexes by either, and
the context helpers (resolve_isec, isec_addr, isec_relocs) keep taking
usize for local arithmetic. Only the places that push into or read out
of an id vector, or index another table by an id, cast - the same split
mold-rust has between its u32 ids and .index().

On the debug clang link this halves ~140MB of id vectors, and an output
symbol table entry drops from 32 to 24 bytes. 322 integration tests
pass; the debug clang, nu and mold links run.

+++ src/icf.rs
@@ -347,7 +347,7 @@ pub fn icf_sections<E: Arch>(ctx: &mut Context<E>) {
                     }
                     return (Edge::Isec(isec, sym.value), addend);
                 }
-                (Edge::Sym(sym_id), addend)
+                (Edge::Sym(sym_id as usize), addend)
             }
             RelocTarget::Section(isec) => {
                 let isec = ctx.resolve_isec(isec as usize);
